requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
notion-client>=2.2.1
openai>=1.12.0
fastapi>=0.104.0
mangum>=0.17.0
supabase>=2.0.0
python-multipart>=0.0.6